fix(toast): separate styles, types and logic to different files
