+++ packages/toaster/src/toast-manager.tsx
@@ -1,35 +1,8 @@
 import * as React from 'react'
-import styled from 'styled-components'
-import {Toast, ToastModel} from './toast'
+import {Toast} from './toast'
+import {ToastModel, ToastManagerProps} from './types'
 import {TransitionGroup} from 'react-transition-group'
-
-const ToastsWrapper = styled.span`
-  max-width: 560;
-  margin: 0 auto;
-  top: 0;
-  left: 0;
-  right: 0;
-  position: fixed;
-  z-index: 30;
-  pointer-events: none;
-`
-
-interface ToastManagerProps {
-  /**
-   * Function called with the `this.notify` function.
-   */
-  bindNotify: any
-
-  /**
-   * Function called with the `this.getToasts` function.
-   */
-  bindGetToasts: any
-
-  /**
-   * Function called with the `this.closeAll` function.
-   */
-  bindCloseAll: any
-}
+import {ToastsWrapper} from './styles'
 
 const hasCustomId = settings => Object.hasOwnProperty.call(settings, 'id')
 
@@ -66,8 +39,8 @@ const ToastManager: React.FC<ToastManagerProps> = props => {
         id,
         title,
         description: settings.description,
-        hasCloseButton: settings.hasCloseButton || true,
-        duration: settings.duration || 5,
+        hasCloseButton: settings.hasCloseButton || false,
+        duration: settings.duration || 0,
         close: () => closeToast(id),
         intent: settings.intent,
         isShown: true
