+++ packages/toaster/src/toast.tsx
@@ -1,91 +1,11 @@
 import * as React from 'react'
-import styled from 'styled-components'
 import {Transition} from 'react-transition-group'
-import {TransitionStatus} from 'react-transition-group/Transition'
 import {useDefaults} from '@smashing/theme'
-import {Alert, AlertAppearanceType} from '@smashing/alert'
+import {Alert} from '@smashing/alert'
+import {ToastProps} from './types'
+import {ANIMATION_DURATION, ToastCloseButton, WrapperAnimated} from './styles'
 
-export interface ToastModel {
-  id: string | number
-  title: string
-  description?: string
-  hasCloseButton: boolean
-  duration: number
-  close: () => void
-  intent: 'success' | 'warning' | 'danger' | 'info'
-  isShown: boolean
-}
-
-const animationEasing = {
-  deceleration: 'cubic-bezier(0.0, 0.0, 0.2, 1)',
-  acceleration: 'cubic-bezier(0.4, 0.0, 1, 1)',
-  spring: 'cubic-bezier(0.175, 0.885, 0.320, 1.175)'
-}
-
-const ANIMATION_DURATION = 200
-
-const WrapperAnimated = styled.div<{state: TransitionStatus}>`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 0;
-  transition: all ${ANIMATION_DURATION}ms ${animationEasing.deceleration};
-
-  ${_ => ({
-    transform: _.state === 'entering' ? 'translateY(-120%)' : 'translateY(0%)',
-    opacity: _.state === 'entered' ? 1 : 0
-  })}
-`
-const ToastCloseButton = styled.span`
-  top: 0;
-  right: 0;
-  position: absolute;
-  pointer-events: all;
-  cursor: pointer;
-  background: yellow;
-  z-index: 31;
-  padding: 10px;
-`
-
-interface Props {
-  /**
-   * The z-index of the toast.
-   */
-  zIndex?: number
-
-  /**
-   * Duration of the toast.
-   */
-  duration?: number
-
-  /**
-   * Function called when the toast is all the way closed.
-   */
-  onRemove?: () => void
-
-  /**
-   * The type of the alert.
-   */
-  intent?: 'success' | 'warning' | 'danger' | 'info'
-
-  /**
-   * The title of the alert.
-   */
-  title?: string
-
-  /**
-   * When true, show a close icon button inside of the toast.
-   */
-  hasCloseButton?: boolean
-
-  /**
-   * When false, will close the Toast and call onRemove when finished.
-   */
-  isShown?: boolean
-  apperance?: AlertAppearanceType
-}
-
-export const Toast: React.FC<Props> = ({children, ...props}) => {
+export const Toast: React.FC<ToastProps> = ({children, ...props}) => {
   const [isShown, setIsShown] = React.useState(true)
   const [closeTimer, setCloseTimer] = React.useState<number | null>(null)
   const [height, setHeight] = React.useState(0)
@@ -151,7 +71,7 @@ export const Toast: React.FC<Props> = ({children, ...props}) => {
 
     setHeight(refHeight)
   }, [])
-  console.dir({defaults})
+
   return (
     <Transition
       in={isShown}
@@ -192,6 +112,6 @@ export const Toast: React.FC<Props> = ({children, ...props}) => {
 declare module 'styled-components' {
   export interface SmashingToastDefaults
     extends Partial<{
-      toaster?: Partial<Props>
+      toaster?: Partial<ToastProps>
     }> {}
 }
